+++ database/migrations/2024_06_04_061727_create_transaction_sale_items_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->string('product_sku');
             $table->string('product_name');
             $table->integer('total_qty');
-            $table->decimal('sale_price', 15, 2);
-            $table->decimal('sale_profit', 15, 2);
+            $table->integer('sale_price');
+            $table->integer('sale_profit');
             $table->string('store_code');
             $table->foreignId('product_id')->nullable();
             $table->foreignId('transaction_id')->references('id')->on('transaction_sales')->cascadeOnDelete();
