+++ database/migrations/2024_07_19_033223_create_store_assets_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('title', 100);
             $table->string('message', 244);
-            $table->decimal('amount', 15, 2);
+            $table->integer('amount');
             $table->enum('type', ['in', 'out', 'hold']);
             $table->string('store_code');
             $table->timestamps();
